Moved files in user-service
