--- backend/service/user-service/src/main/java/com/finance/UserServiceApplication.java
@@ -1,15 +1,0 @@
-package com.finance;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication
-@EntityScan(basePackages = "com.finance.domain")
-@EnableJpaRepositories(basePackages = "com.finance.infrastructure.repository")
-public class UserServiceApplication {
-  public static void main(String[] args) {
-    SpringApplication.run(UserServiceApplication.class, args);
-  }
-}
