--- backend/service/user-service/src/main/java/com/finance/domain/User.java
@@ -1,109 +1,0 @@
-package com.finance.domain;
-
-import com.finance.jwt.security.authentication.UserDetailsExtend;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-@Entity
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "users")
-public class User implements UserDetailsExtend {
-  @Id
-  @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "users_id_sequence")
-  @SequenceGenerator(name = "users_id_sequence", allocationSize = 10)
-  private Long id;
-
-  private String name;
-
-  @Column(unique = true)
-  private String email;
-
-  private String password;
-
-  @Enumerated(EnumType.STRING)
-  private Role role;
-
-  public User(String name, String email, String password) {
-    this.name = name;
-    this.email = email;
-    this.password = password;
-  }
-
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-    return new ArrayList<>(List.of(new SimpleGrantedAuthority(role.getName())));
-  }
-
-  @Override
-  public String getUsername() {
-    return email;
-  }
-
-  @Override
-  public boolean isAccountNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isAccountNonLocked() {
-    return true;
-  }
-
-  @Override
-  public boolean isCredentialsNonExpired() {
-    return true;
-  }
-
-  @Override
-  public boolean isEnabled() {
-    return true;
-  }
-
-  @Override
-  public Long getId() {
-    return id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  @Override
-  public String getPassword() {
-    return password;
-  }
-
-  public Role getRole() {
-    return role;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public void setEmail(String email) {
-    this.email = email;
-  }
-
-  public void setPassword(String password) {
-    this.password = password;
-  }
-
-  public void setRole(Role role) {
-    this.role = role;
-  }
-}
