--- backend/service/user-service/src/main/java/com/finance/infrastructure/log/LoggerConfig.java
@@ -1,9 +1,0 @@
-package com.finance.infrastructure.log;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan("com.finance.logger")
-public class LoggerConfig {
-}
