--- backend/service/user-service/src/main/java/com/finance/infrastructure/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.finance.infrastructure.repository;
-
-
-import com.finance.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-  Optional<User> findUserByEmail(String email);
-}
