--- backend/service/user-service/src/main/java/com/finance/service/UserService.java
@@ -1,11 +1,0 @@
-package com.finance.service;
-
-import com.finance.domain.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
-
-import java.util.List;
-
-public interface UserService extends UserDetailsService {
-  void saveAllIfEmpty(List<User> userList);
-  User save(User user);
-}
