+++ backend/service/user-service/src/main/java/com/finance/user/UserServiceApplication.java
@@ -1,0 +1,15 @@
+package com.finance.user;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication
+@EntityScan(basePackages = "com.finance.domain")
+@EnableJpaRepositories(basePackages = "com.finance.user.infrastructure.repository")
+public class UserServiceApplication {
+  public static void main(String[] args) {
+    SpringApplication.run(UserServiceApplication.class, args);
+  }
+}
