+++ backend/service/user-service/src/main/java/com/finance/user/infrastructure/log/LoggerConfig.java
@@ -1,0 +1,9 @@
+package com.finance.user.infrastructure.log;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@ComponentScan("com.finance.logger")
+public class LoggerConfig {
+}
