+++ backend/service/user-service/src/main/java/com/finance/user/infrastructure/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package com.finance.user.infrastructure.repository;
+
+
+import com.finance.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+  Optional<User> findUserByEmail(String email);
+}
