+++ backend/service/user-service/src/main/java/com/finance/user/service/UserService.java
@@ -1,0 +1,11 @@
+package com.finance.user.service;
+
+import com.finance.user.domain.User;
+import org.springframework.security.core.userdetails.UserDetailsService;
+
+import java.util.List;
+
+public interface UserService extends UserDetailsService {
+  void saveAllIfEmpty(List<User> userList);
+  User save(User user);
+}
