+++ backend/service/user-service/src/main/java/com/finance/user/view/config/ViewConfig.java
@@ -1,0 +1,10 @@
+package com.finance.user.view.config;
+
+import com.finance.exception.config.ExceptionHandlerConfig;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(ExceptionHandlerConfig.class)
+public class ViewConfig {
+}
