+++ backend/service/user-service/src/main/java/com/finance/user/view/dto/LoginDto.java
@@ -1,0 +1,13 @@
+package com.finance.user.view.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class LoginDto {
+  private String email;
+  private String password;
+}
