+++ backend/service/user-service/src/main/java/com/finance/user/view/dto/SignupRequestDto.java
@@ -1,0 +1,14 @@
+package com.finance.user.view.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class SignupRequestDto {
+  private String name;
+  private String email;
+  private String password;
+}
