--- backend/service/user-service/src/main/java/com/finance/view/config/ViewConfig.java
@@ -1,10 +1,0 @@
-package com.finance.view.config;
-
-import com.finance.exception.config.ExceptionHandlerConfig;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-@Configuration
-@Import(ExceptionHandlerConfig.class)
-public class ViewConfig {
-}
