--- backend/service/user-service/src/main/java/com/finance/view/dto/LoginDto.java
@@ -1,13 +1,0 @@
-package com.finance.view.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginDto {
-  private String email;
-  private String password;
-}
