--- backend/service/user-service/src/main/java/com/finance/view/dto/SignupRequestDto.java
@@ -1,14 +1,0 @@
-package com.finance.view.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignupRequestDto {
-  private String name;
-  private String email;
-  private String password;
-}
