--- backend/service/user-service/src/main/java/com/finance/view/dto/SignupResponseDto.java
@@ -1,15 +1,0 @@
-package com.finance.view.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class SignupResponseDto {
-  private Long id;
-  private String name;
-  private String email;
-  private RoleDto role;
-}
