--- backend/service/user-service/src/main/java/com/finance/view/dto/TokenPairDto.java
@@ -1,13 +1,0 @@
-package com.finance.view.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class TokenPairDto {
-  private String access;
-  private String refresh;
-}
